+++ ChangeLog.md
@@ -1,6 +1,7 @@
 ## change log ##
 
 * 2016-06-18: 0.3.2:
+  - extract view-as-options utility into a directive, and also use it in term.html
   - extract clipboard-copy utility into a directive
 
 * 2016-06-17: 0.3.2: